Improve scaling quality.
It is ok to scale with GL_NEAREST on YUV during deinterlace. But when scaling
the RGB converted image to output window size we should use a GL_LINEAR for
better quality.
Therefor we pass the texture filter as parameter to texture creation function
and use the optimal filter depending on the texture we create.