Add quirk to close left opened file handles.
This quirk will run over all file handles of the current process when
we have shutdown EGL and closes all open files handles wich are still
open.
This is reqired because Nvidia's EGL/GLES implementation does not close
the handles when calling eglTerminate. If we would not do this we leaf
8 handles opened each time we change a channel (one for each texture
and shader).