Update build system for proper gst-1.0 support.
While the plugin code itself already supports Gstreamer-1.0 for quite a
while, the build system did not support switching between both API's.

This will add the missing build system support for Gstreamer-0.10 and
Gstreamer 1.0 API. Therefor a new configure option (--with-gstreamer) is
added. As default API Gstreamer-0.10 is selected.

It also splits up the pkg-config check for gstreamer, glesv2 and gio which
was done in one step in the past. This will give us proper error messages,
if one package is missing.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>
Tested-by: Julian Scheel <julian@jusst.de>