Add proper error handling for thread creation.
In some rare cases it can happen that the gl render thread can not be
created. If the thread can not be created the program aborts (refer
g_thread_new documentation). This is not always the best way to go, so
we now handle this error case and creating a flow error.
Note that a flow error expects a message on the bus (see GST_FLOW_ERROR
and GST_ELEMENT_ERROR).

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>